src/requirements.txt
requests
schedule
pynput
pillow